# Make Rounding.nextRoundingValue consistent