--- a/server/src/main/java/org/elasticsearch/common/Rounding.java
+++ b/server/src/main/java/org/elasticsearch/common/Rounding.java
@@ -1157,7 +1157,7 @@ public abstract class Rounding implements Writeable {
                         }
                         continue;
                     }
-                    assert highEnough && false == tooHigh;
+                    assert highEnough && (false == tooHigh);
                     assert roundedRoundedDown == prevRound;
                     if (iterations > 3 && logger.isDebugEnabled()) {
                         logger.debug("Iterated {} time for {} using {}", iterations, utcMillis, TimeIntervalRounding.this.toString());